cleanup + some spice
most of the code previously added to Macroview.tsx has been refactored into a new file LeftAndCenterArea.tsx

- added some animation to the edit panel to make switching between different types of elements less jarring
- adjusted some spacing
